Fix `addError is undefined` issues on Elysia 0.8.9
Upgrade Elysia to `0.8.9` (latest):

```
$ bun pm ls
node_modules (19)
├── bun-types@1.0.24
├── elysia@0.8.9
├── elysia@0.8.9
├── prettier@3.2.4
└── typescript@5.3.3
```

Then fix the breaking change to rename `addError` to `error`. After
doing this there appears to be a bug where the custom error code is no
longer set when the `onError` handler is called (`code` is always
`undefined`); to work around this, use `instanceof` instead to check for
the custom error code.